线稿提取/requirements.txt
torch
importlib_metadata
huggingface_hub
scipy
opencv-python>=4.7.0.72
filelock
numpy
Pillow
einops
torchvision
scikit-image
python-dateutil
scikit-learn
matplotlib
segment_anything
timm
addict
yapf
transformers>=4.39,<5.0
tokenizers>=0.15.2
gradio>=5.20.0
gradio_client>=1.10.0
pydantic>=2.11.1